+++ dandiapi/api/tests/test_tasks.py
@@ -1,15 +1,11 @@
 import hashlib
 
-from dandischema.digests.dandietag import mb
 from django.conf import settings
 from django.core.files.storage import Storage
-import factory
 import pytest
 
 from dandiapi.api import tasks
-from dandiapi.api.copy import CopyPartGenerator
 from dandiapi.api.models import Asset, AssetBlob, EmbargoedAssetBlob, Version
-from dandiapi.api.models.dandiset import Dandiset
 
 
 @pytest.mark.django_db
@@ -254,171 +250,3 @@ def test_validate_version_metadata_malformed_license(version: Version, asset: As
     assert version.validation_errors == [
         {'field': 'license', 'message': "'foo' is not of type 'array'"}
     ]
-
-
-@pytest.mark.django_db
-def test_unembargo_dandiset(
-    dandiset_factory,
-    draft_version_factory,
-    asset_factory,
-    embargoed_asset_blob_factory,
-    storage,
-    embargoed_storage,
-    monkeypatch,
-):
-    # Since both fixtures are parametrized, only proceed when they use the same storage backend
-    if type(storage) != type(embargoed_storage):
-        pytest.skip('Skip tests with mismatched storages')
-
-    # Monkey patch the part size in CopyPartGenerator, so it can be tested with smaller part sizes
-    monkeypatch.setattr(CopyPartGenerator, 'DEFAULT_PART_SIZE', mb(6), raising=True)
-
-    # Pretend like AssetBlob/EmbargoedAssetBlob were defined with the given storage
-    AssetBlob.blob.field.storage = storage
-    EmbargoedAssetBlob.blob.field.storage = embargoed_storage
-
-    # Create dandiset and version
-    dandiset: Dandiset = dandiset_factory(embargo_status=Dandiset.EmbargoStatus.EMBARGOED)
-    draft_version: Version = draft_version_factory(dandiset=dandiset)
-
-    # Create an embargoed asset that's 10mb in size
-    embargoed_asset_blob: EmbargoedAssetBlob = embargoed_asset_blob_factory(size=mb(10))
-    embargoed_asset: Asset = asset_factory(embargoed_blob=embargoed_asset_blob, blob=None)
-    draft_version.assets.add(embargoed_asset)
-
-    # Assert properties before unembargo
-    assert embargoed_asset.embargoed_blob is not None
-    assert embargoed_asset.blob is None
-    assert embargoed_asset.embargoed_blob.etag != ''
-
-    # Run unembargo
-    tasks.unembargo_dandiset(dandiset.pk)
-    dandiset.refresh_from_db()
-    draft_version.refresh_from_db()
-
-    # Assert correct changes took place
-    assert dandiset.embargo_status == Dandiset.EmbargoStatus.OPEN
-    assert draft_version.metadata['access'] == 'dandi:OpenAccess'
-
-    # Assert no new asset created
-    asset: Asset = draft_version.assets.first()
-    assert asset == embargoed_asset
-
-    # Check blobs
-    assert asset.embargoed_blob is None
-    assert asset.blob is not None
-    assert asset.blob.etag == embargoed_asset_blob.etag
-
-    blob_id = str(asset.blob.blob_id)
-    assert asset.blob.blob.name == f'test-prefix/blobs/{blob_id[:3]}/{blob_id[3:6]}/{blob_id}'
-
-
-@pytest.mark.django_db
-def test_unembargo_dandiset_existing_blobs(
-    dandiset_factory,
-    draft_version_factory,
-    asset_factory,
-    asset_blob_factory,
-    embargoed_asset_blob_factory,
-    storage,
-    embargoed_storage,
-):
-    # Since both fixtures are parametrized, only proceed when they use the same storage backend
-    if type(storage) != type(embargoed_storage):
-        pytest.skip('Skip tests with mismatched storages')
-
-    # Pretend like AssetBlob/EmbargoedAssetBlob were defined with the given storage
-    AssetBlob.blob.field.storage = storage
-    EmbargoedAssetBlob.blob.field.storage = embargoed_storage
-
-    # Create dandiset and version
-    dandiset: Dandiset = dandiset_factory(embargo_status=Dandiset.EmbargoStatus.EMBARGOED)
-    draft_version: Version = draft_version_factory(dandiset=dandiset)
-
-    # Create embargoed assets
-    embargoed_asset_blob: EmbargoedAssetBlob = embargoed_asset_blob_factory()
-    embargoed_asset: Asset = asset_factory(embargoed_blob=embargoed_asset_blob, blob=None)
-    draft_version.assets.add(embargoed_asset)
-
-    # Create unembargoed asset with identical data
-    embargoed_asset_blob_data = embargoed_asset_blob.blob.read()
-    embargoed_asset_blob.blob.seek(0)
-    existing_asset_blob = asset_blob_factory(
-        blob=factory.django.FileField(data=embargoed_asset_blob_data)
-    )
-
-    # Assert properties before unembargo
-    assert embargoed_asset.embargoed_blob is not None
-    assert embargoed_asset.blob is None
-    assert embargoed_asset.embargoed_blob.etag != ''
-    assert existing_asset_blob.etag != ''
-    assert embargoed_asset_blob.etag == existing_asset_blob.etag
-
-    # Run unembargo
-    tasks.unembargo_dandiset(dandiset.pk)
-    dandiset.refresh_from_db()
-    draft_version.refresh_from_db()
-
-    # Assert correct changes took place
-    assert dandiset.embargo_status == Dandiset.EmbargoStatus.OPEN
-    assert draft_version.metadata['access'] == 'dandi:OpenAccess'
-
-    # Assert no new asset created
-    asset: Asset = draft_version.assets.first()
-    assert asset == embargoed_asset
-
-    # Check blobs
-    assert asset.embargoed_blob is None
-    assert asset.blob is not None
-    assert asset.blob.etag == embargoed_asset_blob.etag
-    assert asset.blob == existing_asset_blob
-
-
-@pytest.mark.django_db
-def test_unembargo_dandiset_normal_asset_blob(
-    dandiset_factory,
-    draft_version_factory,
-    asset_factory,
-    asset_blob_factory,
-    storage,
-    embargoed_storage,
-):
-    # Since both fixtures are parametrized, only proceed when they use the same storage backend
-    if type(storage) != type(embargoed_storage):
-        pytest.skip('Skip tests with mismatched storages')
-
-    # Pretend like AssetBlob was defined with the given storage
-    AssetBlob.blob.field.storage = storage
-
-    # Create dandiset and version
-    dandiset: Dandiset = dandiset_factory(embargo_status=Dandiset.EmbargoStatus.EMBARGOED)
-    draft_version: Version = draft_version_factory(dandiset=dandiset)
-
-    # Create asset
-    asset_blob: AssetBlob = asset_blob_factory()
-    asset: Asset = asset_factory(blob=asset_blob, embargoed_blob=None)
-    draft_version.assets.add(asset)
-
-    # Assert properties before unembargo
-    assert asset.embargoed_blob is None
-    assert asset.blob is not None
-
-    # Run unembargo
-    tasks.unembargo_dandiset(dandiset.pk)
-    dandiset.refresh_from_db()
-    draft_version.refresh_from_db()
-
-    # Assert correct changes took place
-    assert dandiset.embargo_status == Dandiset.EmbargoStatus.OPEN
-    assert draft_version.metadata['access'] == 'dandi:OpenAccess'
-
-    # Assert no new asset created
-    fetched_asset: Asset = draft_version.assets.first()
-    assert asset == fetched_asset
-
-    # Check that blob is unchanged
-    assert fetched_asset.blob == asset_blob
-    assert asset.embargoed_blob is None
-    assert asset.blob is not None
-    assert asset.blob.etag
-    assert asset.blob
